sme_plugin/requirements.txt
langchain
langchain-google-genai
langchain-community
pypdf
watchdog
pyyaml
